Base list of extensions
This is a list I have put together. I may go back and do a complete wipe of the gold master because before I took this snapshot I started doing some work on the machine and added some of my extras. Take a look, if you want to add items that would be great.